requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.10
fastapi>=0.110.0
Jinja2>=3.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
requests>=2.31.0
uvicorn[standard]>=0.27.0
